cli/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "infragpt"
version = "0.1.52"
description = "InfraGPT is an AI assistant for interactively managing your infrastructure."
readme = "README.md"
authors = [{ name = "Priyanshu Jain", email = "pj@73ai.org" }]
requires-python = ">=3.13"
license = { text = "GPL-3.0-only" }
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
    "Topic :: Utilities",
    "Topic :: System :: Systems Administration",
]

dependencies = [
    "rich>=14.0.0",
    "prompt-toolkit>=3.0.51",
    "click>=8.2.1",
    "pyperclip>=1.9.0",
    "pyyaml>=6.0.2",
    "openai>=1.98.0",
    "anthropic>=0.60.0",
    "docker>=7.1.0",
    "cryptography>=44.0.1",
    "httpx>=0.28.1",
]

[project.urls]
Homepage = "https://github.com/73ai/infragpt"

[project.scripts]
infragpt = "infragpt.main:cli"




[dependency-groups]
dev = [
    "pyright>=1.1.407",
    "pytest>=8.4.1",
    "ruff>=0.14.10",
]
